完成三期基本需求

+++ systemServe/tornado_serve/person/cancel_foucs.py
@@ -6,7 +6,7 @@ sys.path.append("..")
 from judge_permission import judgeIfPermiss
 
 from api_define import cancel_focus
-from orm import *
+from tornado_serve.orm import db,stu_focus,stu_basic_info
 
 class CancelFocus(cancel_focus):
 
@@ -29,9 +29,13 @@ class CancelFocus(cancel_focus):
         """
         try:
             with db.execution_context():
-                data = {"state": 0,'focusColor':'blue'}
-                stu_focus.delete().where(stu_focus.stuID == stu_id).execute()
-                stu_basic_info.update(**data).where(stu_basic_info.stuID == stu_id).execute()
+                thisStu=stu_basic_info.select(stu_basic_info.state).where(stu_basic_info.stuID==stu_id).get()
+                if thisStu.state==1:
+                    return {"status": 1, "errorInfo": "该学生处于学情关注状态，若需要取消关注状态，请在学生的学情个人页进行操作"}
+                else:
+                    data = {"state": 0,'focusColor':'blue'}
+                    stu_focus.delete().where(stu_focus.stuID == stu_id).execute()
+                    stu_basic_info.update(**data).where(stu_basic_info.stuID == stu_id).execute()
         except:
             raise
             # return {"status":0, "errorInfo":"数据库新增信息失败，请稍候重试"}
